updating the rotations
Underneath the rotations class is updated. this may break unexpected
uses of Euler angles because the rotations for azimuth and tilt were
swapped in the math in the original API. This update fixes the
fundamental flaw in the rotations library that failed whenever azimuth
was greater than 90 or less than -90.