Fixing the regeneration problem with pasted items
Boils down to making sure that all added objects are added to the
creaated names, but then are filtered out in the recoursive apply just
like the locked items. THis allows the parameterics to be added to the
new items, but also prevents then additional items from being selected
in the selection box.